fix(e2e): strip markdown in saved-reminder push + centralize encrypted label
Addresses the two CodeRabbit findings on the saved-reminder preview:

- INV-60: `resolveSavedReminderPreview` sliced raw `contentMarkdown` into the
  push body, so `**bold**`/`[text](url)` shipped as literal syntax in the OS
  notification (the service worker renders the preview verbatim — there is no
  later strip step). Route non-placeholder content through `stripMarkdownToInline`
  before truncating. The pure strip impl moves to `@threa/types` so the backend
  can share it; the frontend `lib/markdown/strip` re-exports it (one impl, all
  existing importers unchanged).
- INV-33: the "Encrypted message" label was hardcoded in three surfaces (push,
  Saved list, sidebar). Centralize it as `ENCRYPTED_MESSAGE_PREVIEW_LABEL` next
  to `E2E_PLACEHOLDER_CONTENT_MARKDOWN` and point all three at it so they can't
  drift.

Tests: push preview unit suite (incl. new markdown-strip case), frontend strip +
saved-item via vitest, monorepo typecheck, backend + frontend eslint — all green.
