+++ apps/backend/src/features/push/service.ts
@@ -9,6 +9,8 @@ import {
   StreamTypes,
   PRESENCE_INTERACTION_WINDOW_MS,
   E2E_PLACEHOLDER_CONTENT_MARKDOWN,
+  ENCRYPTED_MESSAGE_PREVIEW_LABEL,
+  stripMarkdownToInline,
   type PrefNotificationLevel,
   type StreamType,
 } from "@threa/types"
@@ -93,11 +95,14 @@ interface PushServiceDeps {
  * zero-width placeholder on the wire (the server holds no key), so surfacing
  * the raw markdown produces a blank notification — substitute a generic,
  * leak-free label instead (E2EE-19), mirroring the saved-list/sidebar
- * treatment. Normal messages get their body truncated; missing content → null.
+ * treatment. Normal messages are stripped to plain text before truncation so
+ * the OS notification never shows literal markdown syntax (INV-60 — the SW
+ * renders this verbatim, there is no later strip step). Missing content → null.
  */
 export function resolveSavedReminderPreview(contentMarkdown: string | null | undefined): string | null {
-  if (contentMarkdown === E2E_PLACEHOLDER_CONTENT_MARKDOWN) return "🔒 Encrypted message"
-  return contentMarkdown?.slice(0, 200) ?? null
+  if (contentMarkdown === E2E_PLACEHOLDER_CONTENT_MARKDOWN) return ENCRYPTED_MESSAGE_PREVIEW_LABEL
+  if (contentMarkdown == null) return null
+  return stripMarkdownToInline(contentMarkdown).slice(0, 200)
 }
 
 export class PushService {
