+++ apps/frontend/src/components/saved/saved-item.tsx
@@ -1,7 +1,12 @@
 import { useState } from "react"
 import { Link } from "react-router-dom"
 import { Archive, Bell, Check, CircleAlert, Trash2, Undo2 } from "lucide-react"
-import { E2E_PLACEHOLDER_CONTENT_MARKDOWN, type SavedMessageView, type SavedStatus } from "@threa/types"
+import {
+  E2E_PLACEHOLDER_CONTENT_MARKDOWN,
+  ENCRYPTED_MESSAGE_PREVIEW_LABEL,
+  type SavedMessageView,
+  type SavedStatus,
+} from "@threa/types"
 import { Button } from "@/components/ui/button"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
 import { cn } from "@/lib/utils"
@@ -195,7 +200,7 @@ function resolvePreview(saved: SavedMessageView): string {
     // ciphertext to the Saved surface, so there's nothing to decrypt here —
     // render the same "Encrypted message" treatment the sidebar uses (E2EE-16)
     // rather than a blank/invisible preview.
-    if (saved.message.contentMarkdown === E2E_PLACEHOLDER_CONTENT_MARKDOWN) return "🔒 Encrypted message"
+    if (saved.message.contentMarkdown === E2E_PLACEHOLDER_CONTENT_MARKDOWN) return ENCRYPTED_MESSAGE_PREVIEW_LABEL
     return stripMarkdownToInline(saved.message.contentMarkdown)
   }
   if (saved.unavailableReason === "deleted") return "Original message was deleted"
